made separate file for quick alarm viewmodel
